Backend/claude-langchain-bot/requirements.txt
fastapi
uvicorn
langchain-anthropic
langchain-core
python-dotenv
